requirements.txt
docx==0.2.4
faiss_cpu==1.12.0
fastapi==0.121.1
fitz==0.0.1.dev2
keybert==0.9.0
numpy==2.3.4
python_docx==1.2.0
PyYAML==6.0.3
Requests==2.32.5
sentence_transformers==5.1.1
streamlit==1.51.0
tqdm==4.67.1
transformers==4.56.2
uvicorn==0.38.0
Hf_xet